+++ src/contract/default.ts
@@ -1,4 +1,5 @@
-import type { Abi as AbiKanabi } from 'abi-wan-kanabi';
+import type { Abi as AbiKanabiV1 } from 'abi-wan-kanabi-v1';
+import type { Abi as AbiKanabiV2, TypedContract as AbiWanTypedContractV2 } from 'abi-wan-kanabi-v2';
 
 import { AccountInterface } from '../account';
 import { ProviderInterface, defaultProvider } from '../provider';
@@ -30,7 +31,9 @@ import { CallData, cairo } from '../utils/calldata';
 import { createAbiParser } from '../utils/calldata/parser';
 import { getAbiEvents, parseEvents as parseRawEvents } from '../utils/events/index';
 import { cleanHex } from '../utils/num';
-import { ContractInterface, TypedContract } from './interface';
+import { ContractInterface, TypedContractV1 } from './interface';
+
+export type TypedContractV2<TAbi extends AbiKanabiV2> = AbiWanTypedContractV2<TAbi> & Contract;
 
 export const splitArgsAndOptions = (args: ArgsOrCalldataWithOptions) => {
   const options = [
@@ -346,7 +349,11 @@ export class Contract implements ContractInterface {
     return this.providerOrAccount.getContractVersion(this.address);
   }
 
-  public typed<TAbi extends AbiKanabi>(tAbi: TAbi): TypedContract<TAbi> {
-    return this as TypedContract<typeof tAbi>;
+  public typedv1<TAbi extends AbiKanabiV1>(tAbi: TAbi): TypedContractV1<TAbi> {
+    return this as TypedContractV1<typeof tAbi>;
+  }
+
+  public typedv2<TAbi extends AbiKanabiV2>(tAbi: TAbi): TypedContractV2<TAbi> {
+    return this as unknown as TypedContractV2<typeof tAbi>;
   }
 }