+++ src/utils/typedData.ts
@@ -1,11 +1,96 @@
-import { BigNumberish, StarkNetMerkleType, StarkNetType, TypedData } from '../types';
-import { computeHashOnElements, getSelectorFromName } from './hash';
+/* eslint-disable no-param-reassign */
+import {
+  BigNumberish,
+  TypedDataRevision as Revision,
+  StarkNetEnumType,
+  StarkNetMerkleType,
+  StarkNetType,
+  TypedData,
+} from '../types';
+import {
+  computePedersenHash,
+  computePedersenHashOnElements,
+  computePoseidonHash,
+  computePoseidonHashOnElements,
+  getSelectorFromName,
+} from './hash';
 import { MerkleTree } from './merkle';
 import { isHex, toHex } from './num';
-import { encodeShortString } from './shortString';
+import { encodeShortString, splitLongString } from './shortString';
 
 /** @deprecated prefer importing from 'types' over 'typedData' */
 export * from '../types/typedData';
+
+interface Context {
+  parent?: string;
+  key?: string;
+}
+
+interface Configuration {
+  domain: string;
+  hashMethod: (data: BigNumberish[]) => string;
+  hashMerkleMethod: (a: BigNumberish, b: BigNumberish) => string;
+  escapeTypeString: (s: string) => string;
+  presetTypes: TypedData['types'];
+}
+
+const presetTypes: TypedData['types'] = {
+  u256: JSON.parse('[{ "name": "low", "type": "u128" }, { "name": "high", "type": "u128" }]'),
+  TokenAmount: JSON.parse(
+    '[{ "name": "token_address", "type": "ContractAddress" }, { "name": "amount", "type": "u256" }]'
+  ),
+  NftId: JSON.parse(
+    '[{ "name": "collection_address", "type": "ContractAddress" }, { "name": "token_id", "type": "u256" }]'
+  ),
+};
+
+const revisionConfiguration: Record<Revision, Configuration> = {
+  [Revision.Active]: {
+    domain: 'StarknetDomain',
+    hashMethod: computePoseidonHashOnElements,
+    hashMerkleMethod: computePoseidonHash,
+    escapeTypeString: (s) => `"${s}"`,
+    presetTypes,
+  },
+  [Revision.Legacy]: {
+    domain: 'StarkNetDomain',
+    hashMethod: computePedersenHashOnElements,
+    hashMerkleMethod: computePedersenHash,
+    escapeTypeString: (s) => s,
+    presetTypes: {},
+  },
+};
+
+// TODO: replace with utils byteArrayFromString from PR#891 once it is available
+export function byteArrayFromString(targetString: string) {
+  const shortStrings: string[] = splitLongString(targetString);
+  const remainder: string = shortStrings[shortStrings.length - 1];
+  const shortStringsEncoded: BigNumberish[] = shortStrings.map(encodeShortString);
+
+  const [pendingWord, pendingWordLength] =
+    remainder === undefined || remainder.length === 31
+      ? ['0x00', 0]
+      : [shortStringsEncoded.pop()!, remainder.length];
+
+  return {
+    data: shortStringsEncoded.length === 0 ? ['0x00'] : shortStringsEncoded,
+    pending_word: pendingWord,
+    pending_word_len: pendingWordLength,
+  };
+}
+
+function identifyRevision({ types, domain }: TypedData) {
+  if (revisionConfiguration[Revision.Active].domain in types && domain.revision === Revision.Active)
+    return Revision.Active;
+
+  if (
+    revisionConfiguration[Revision.Legacy].domain in types &&
+    (domain.revision ?? Revision.Legacy) === Revision.Legacy
+  )
+    return Revision.Legacy;
+
+  return undefined;
+}
 
 function getHex(value: BigNumberish): string {
   try {
@@ -20,18 +105,13 @@ function getHex(value: BigNumberish): string {
 
 /**
  * Validates that `data` matches the EIP-712 JSON schema.
- *
- * @param {any} data
- * @return {boolean}
  */
-const validateTypedData = (data: unknown): data is TypedData => {
+function validateTypedData(data: unknown): data is TypedData {
   const typedData = data as TypedData;
-
-  // Validate that the data matches the EIP-712 JSON schema
-  const valid = Boolean(typedData.types && typedData.primaryType && typedData.message);
-
-  return valid;
-};
+  return Boolean(
+    typedData.message && typedData.primaryType && typedData.types && identifyRevision(typedData)
+  );
+}
 
 export function prepareSelector(selector: string): string {
   return isHex(selector) ? selector : getSelectorFromName(selector);
@@ -41,52 +121,48 @@ export function isMerkleTreeType(type: StarkNetType): type is StarkNetMerkleType
   return type.type === 'merkletree';
 }
 
-interface Context {
-  parent?: string;
-  key?: string;
-}
-
 /**
  * Get the dependencies of a struct type. If a struct has the same dependency multiple times, it's only included once
  * in the resulting array.
- *
- * @param {TypedData} typedData
- * @param {string} type
- * @param {string[]} [dependencies]
- * @return {string[]}
  */
-export const getDependencies = (
+export function getDependencies(
   types: TypedData['types'],
   type: string,
-  dependencies: string[] = []
-): string[] => {
+  dependencies: string[] = [],
+  contains: string = '',
+  revision: Revision = Revision.Legacy
+): string[] {
   // Include pointers (struct arrays)
   if (type[type.length - 1] === '*') {
-    // eslint-disable-next-line no-param-reassign
     type = type.slice(0, -1);
+  } else if (revision === Revision.Active) {
+    // enum base
+    if (type === 'enum') {
+      type = contains;
+    }
+    // enum element types
+    else if (type.match(/^\(.*\)$/)) {
+      type = type.slice(1, -1);
+    }
   }
 
-  if (dependencies.includes(type)) {
-    return dependencies;
-  }
-
-  if (!types[type]) {
+  if (dependencies.includes(type) || !types[type]) {
     return dependencies;
   }
 
   return [
     type,
-    ...types[type].reduce<string[]>(
+    ...(types[type] as StarkNetEnumType[]).reduce<string[]>(
       (previous, t) => [
         ...previous,
-        ...getDependencies(types, t.type, previous).filter(
+        ...getDependencies(types, t.type, previous, t.contains, revision).filter(
           (dependency) => !previous.includes(dependency)
         ),
       ],
       []
     ),
   ];
-};
+}
 
 function getMerkleTreeType(types: TypedData['types'], ctx: Context) {
   if (ctx.parent && ctx.key) {
@@ -107,130 +183,215 @@ function getMerkleTreeType(types: TypedData['types'], ctx: Context) {
 /**
  * Encode a type to a string. All dependent types are alphabetically sorted.
  */
-export const encodeType = (types: TypedData['types'], type: string): string => {
-  const [primary, ...dependencies] = getDependencies(types, type);
+export function encodeType(
+  types: TypedData['types'],
+  type: string,
+  revision: Revision = Revision.Legacy
+): string {
+  const [primary, ...dependencies] = getDependencies(types, type, undefined, undefined, revision);
   const newTypes = !primary ? [] : [primary, ...dependencies.sort()];
+
+  const esc = revisionConfiguration[revision].escapeTypeString;
 
   return newTypes
     .map((dependency) => {
-      return `${dependency}(${types[dependency].map((t) => `${t.name}:${t.type}`)})`;
+      const dependencyElements = types[dependency].map((t) => {
+        const targetType =
+          t.type === 'enum' && revision === Revision.Active
+            ? (t as StarkNetEnumType).contains
+            : t.type;
+        // parentheses handling for enum variant types
+        const typeString = targetType.match(/^\(.*\)$/)
+          ? `(${targetType
+              .slice(1, -1)
+              .split(',')
+              .map((e) => (e ? esc(e) : e))
+              .join(',')})`
+          : esc(targetType);
+        return `${esc(t.name)}:${typeString}`;
+      });
+      return `${esc(dependency)}(${dependencyElements})`;
     })
     .join('');
-};
+}
 
 /**
  * Get a type string as hash.
  */
-export const getTypeHash = (types: TypedData['types'], type: string): string => {
-  return getSelectorFromName(encodeType(types, type));
-};
+export function getTypeHash(
+  types: TypedData['types'],
+  type: string,
+  revision: Revision = Revision.Legacy
+): string {
+  return getSelectorFromName(encodeType(types, type, revision));
+}
 
 /**
  * Encodes a single value to an ABI serialisable string, number or Buffer. Returns the data as tuple, which consists of
  * an array of ABI compatible types, and an array of corresponding values.
  */
-export const encodeValue = (
+export function encodeValue(
   types: TypedData['types'],
   type: string,
   data: unknown,
-  ctx: Context = {}
-): [string, string] => {
+  ctx: Context = {},
+  revision: Revision = Revision.Legacy
+): [string, string] {
   if (types[type]) {
-    // eslint-disable-next-line @typescript-eslint/no-use-before-define
-    return [type, getStructHash(types, type, data as Record<string, unknown>)];
+    return [type, getStructHash(types, type, data as TypedData['message'], revision)];
   }
 
-  if (
-    Object.keys(types)
-      .map((x) => `${x}*`)
-      .includes(type)
-  ) {
-    const structHashes: string[] = (data as unknown[]).map((struct) => {
-      // eslint-disable-next-line @typescript-eslint/no-use-before-define
-      return getStructHash(types, type.slice(0, -1), struct as Record<string, unknown>);
-    });
-    return [type, computeHashOnElements(structHashes)];
+  if (revisionConfiguration[revision].presetTypes[type]) {
+    return [
+      type,
+      getStructHash(
+        revisionConfiguration[revision].presetTypes,
+        type,
+        data as TypedData['message'],
+        revision
+      ),
+    ];
   }
 
-  if (type === 'merkletree') {
-    const merkleTreeType = getMerkleTreeType(types, ctx);
-    const structHashes: string[] = (data as unknown[]).map((struct) => {
-      // eslint-disable-next-line @typescript-eslint/no-use-before-define
-      return encodeValue(types, merkleTreeType, struct as Record<string, unknown>)[1];
-    });
-    const { root } = new MerkleTree(structHashes as string[]);
-    return ['felt', root];
+  if (type.endsWith('*')) {
+    const hashes: string[] = (data as Array<TypedData['message']>).map(
+      (entry) => encodeValue(types, type.slice(0, -1), entry, undefined, revision)[1]
+    );
+    return [type, revisionConfiguration[revision].hashMethod(hashes)];
   }
 
-  if (type === 'felt*') {
-    return ['felt*', computeHashOnElements(data as string[])];
-  }
+  switch (type) {
+    case 'enum': {
+      if (revision === Revision.Active) {
+        const [variantKey, variantData] = Object.entries(data as TypedData['message'])[0];
 
-  if (type === 'selector') {
-    return ['felt', prepareSelector(data as string)];
-  }
+        const parentType = types[ctx.parent as string][0] as StarkNetEnumType;
+        const enumType = types[parentType.contains];
+        const variantType = enumType.find((t) => t.name === variantKey) as StarkNetType;
+        const variantIndex = enumType.indexOf(variantType);
 
-  return [type, getHex(data as string)];
-};
+        const encodedSubtypes = variantType.type
+          .slice(1, -1)
+          .split(',')
+          .map((subtype, index) => {
+            if (!subtype) return subtype;
+            const subtypeData = (variantData as unknown[])[index];
+            return encodeValue(types, subtype, subtypeData, undefined, revision)[1];
+          });
+        return [
+          type,
+          revisionConfiguration[revision].hashMethod([variantIndex, ...encodedSubtypes]),
+        ];
+      } // else fall through to default
+      return [type, getHex(data as string)];
+    }
+    case 'merkletree': {
+      const merkleTreeType = getMerkleTreeType(types, ctx);
+      const structHashes: string[] = (data as Array<TypedData['message']>).map((struct) => {
+        return encodeValue(types, merkleTreeType, struct, undefined, revision)[1];
+      });
+      const { root } = new MerkleTree(
+        structHashes as string[],
+        revisionConfiguration[revision].hashMerkleMethod
+      );
+      return ['felt', root];
+    }
+    case 'selector': {
+      return ['felt', prepareSelector(data as string)];
+    }
+    case 'string': {
+      if (revision === Revision.Active) {
+        const byteArray = byteArrayFromString(data as string);
+        const elements = [
+          byteArray.data.length,
+          ...byteArray.data,
+          byteArray.pending_word,
+          byteArray.pending_word_len,
+        ];
+        return [type, revisionConfiguration[revision].hashMethod(elements)];
+      } // else fall through to default
+      return [type, getHex(data as string)];
+    }
+    case 'felt':
+    case 'bool':
+    case 'u128':
+    case 'i128':
+    case 'ContractAddress':
+    case 'ClassHash':
+    case 'timestamp':
+    case 'shortstring':
+      return [type, getHex(data as string)];
+    default: {
+      if (revision === Revision.Active) {
+        throw new Error(`Unsupported type: ${type}`);
+      }
+      return [type, getHex(data as string)];
+    }
+  }
+}
 
 /**
  * Encode the data to an ABI encoded Buffer. The data should be a key -> value object with all the required values.
  * All dependent types are automatically encoded.
  */
-export const encodeData = <T extends TypedData>(
+export function encodeData<T extends TypedData>(
   types: T['types'],
   type: string,
-  data: T['message']
-) => {
-  const [returnTypes, values] = types[type].reduce<[string[], string[]]>(
+  data: T['message'],
+  revision: Revision = Revision.Legacy
+) {
+  const targetType = types[type] ?? revisionConfiguration[revision].presetTypes[type];
+  const [returnTypes, values] = targetType.reduce<[string[], string[]]>(
     ([ts, vs], field) => {
-      if (data[field.name] === undefined || data[field.name] === null) {
+      if (data[field.name] === undefined || (data[field.name] === null && field.type !== 'enum')) {
         throw new Error(`Cannot encode data: missing data for '${field.name}'`);
       }
 
       const value = data[field.name];
-      const [t, encodedValue] = encodeValue(types, field.type, value, {
-        parent: type,
-        key: field.name,
-      });
+      const ctx = { parent: type, key: field.name };
+      const [t, encodedValue] = encodeValue(types, field.type, value, ctx, revision);
 
       return [
         [...ts, t],
         [...vs, encodedValue],
       ];
     },
-    [['felt'], [getTypeHash(types, type)]]
+    [['felt'], [getTypeHash(types, type, revision)]]
   );
 
   return [returnTypes, values];
-};
+}
 
 /**
  * Get encoded data as a hash. The data should be a key -> value object with all the required values.
  * All dependent types are automatically encoded.
  */
-export const getStructHash = <T extends TypedData>(
+export function getStructHash<T extends TypedData>(
   types: T['types'],
   type: string,
-  data: T['message']
-) => {
-  return computeHashOnElements(encodeData(types, type, data)[1]);
-};
+  data: T['message'],
+  revision: Revision = Revision.Legacy
+) {
+  return revisionConfiguration[revision].hashMethod(encodeData(types, type, data, revision)[1]);
+}
 
 /**
- * Get the EIP-191 encoded message to sign, from the typedData object.
+ * Get the SNIP-12 encoded message to sign, from the typedData object.
  */
-export const getMessageHash = (typedData: TypedData, account: BigNumberish): string => {
+export function getMessageHash(typedData: TypedData, account: BigNumberish): string {
   if (!validateTypedData(typedData)) {
     throw new Error('Typed data does not match JSON schema');
   }
 
+  const revision = identifyRevision(typedData) as Revision;
+  const { domain, hashMethod } = revisionConfiguration[revision];
+
   const message = [
     encodeShortString('StarkNet Message'),
-    getStructHash(typedData.types, 'StarkNetDomain', typedData.domain),
+    getStructHash(typedData.types, domain, typedData.domain, revision),
     account,
-    getStructHash(typedData.types, typedData.primaryType, typedData.message),
+    getStructHash(typedData.types, typedData.primaryType, typedData.message, revision),
   ];
 
-  return computeHashOnElements(message);
-};
+  return hashMethod(message);
+}